+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to the "AL-Navigator" extension will be documented in this f
 ## Release Notes
 ## [0.4.0]
 - Fix indentation when creating variables with CodeAction
-
+- New Feature: Create Report (copies a standard report including layout)
 ## [0.3.9]
 - New feature: Open Microsoft translation target language -> english language
 - New feature: Show Microsoft translation target language -> english language